Create `PinghApplication`.
